backend/requirements.txt
asgiref==3.5.2
cffi==1.15.1
Django==4.1.4
django-colorfield==0.8.0
djangorestframework==3.14.0
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
Pillow==9.4.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.7
zope.event==4.6
zope.interface==5.5.2
